server: fix race condition to help testing
Previously, the normal python's Event could be sent by producers,
but not noticed by the manager's thread.

This a really tiny race condition between 'wait() and clear()'.
Between those two calls, some producer allocator/terminator could
set the event, but manager wouldn't notice.  In such cases, the
overall manager's loop would wait till the timeout.

As such, this was not a disaster, but anyways it made the server
less responsive.