manager: use thread-local data where appropriate
It happened before that two TerminateWorker threads were started
with improperly used global storage for object's data.  This led
to the situation that the faster thread got pool_id/resource_name
overwritten by slower thread, leading to the situation that both
threads were deleting the same resource (which in turn means that
one resource was not deleted at all).

Fixes: #19