server: allow re-using allocated resources
This is quite big change, but is compatible with both (a) previous
clients and (b) previous database layout (valid alembic migration path
up is provided).

This patch makes the model actually more robust, because the reference
from resource to ticket is unique, and easily accessible (FK).

Since the patch caused headaches with testsuite in Travis, the overall
output from the testsuite is a bit more verbose now.