Cleanup broken instances after server restart
As long as the configured cmd_* scripting is "atomic", we may be sure
that we don't leave any leftover releases in inconsistent state.  But
when we **restart** the server when it is executing something by any
background worker, the worker gets killed as well as the server and the
resource might never be flipped back to a "consistent" UP state.

Since the resources are meant to disposable, and it should be OK to
re-run the cmd_delete command multiple times - rather request a
termination of all the inconsistent resources when server starts.  It
would be nice to have nicer solution to #41 to avoid resource wasting,
but this patch fixes the most problems.

Relates: #41