server: minimize the use of globals
Like 'session', 'engine', 'db', etc.  Global variables complicate our
testing because the code evaluation is import-order dependant (globals
are initialized when imports are done).  So when testing - it is e.g.
pretty complicated to set "CONFIG_DIR" env variable to a temporary
directory per each test-case as it needs to be set before the config
import is done.