server: more verbose RELEASING thread
Sometimes workers get stuck in RELEASING state forever (until server
restart, when all the RELEASING resources are removed).  See the
Copr issue #2115 for more info [1].

Some related output in hook.log exists, but because the affected
resources usually get into RELEASING state times during their lifetimes,
it's not easy to tell if
(a) the failed ReleaseWorker run for the affected resource actually
failed, or
(b) if it started, but the cmd_release command (Popen) failed, or
(c) ReleaseWorker only failed to store the release status to DB.
With this commit, we should have an evidence next time and see what is
the problem to be fixed.

Relates: https://pagure.io/copr/copr/issue/2115