More intransigent control for failed resources
Fixes: https://pagure.io/copr/copr/issue/2083

Sometimes it happens that even though the resource has check_failed_count set to a number greater than 3, it is not deleted because it still has a ticket. This leads to endless jobs. We can be more intransigent and remove the resource after 10 checks, even though it still has a ticket.