webui: implement /pools page
It returns the same values as the /resources except for not counting
the ENDED resources because of performance reasons.

I believe this implementation fixes the semantic errors of the
previous one, and also evaluates immediately even with a large
database.

The reason why I am adding /pools instead of fixing /resources is
because it was IMHO incorrectly named. The page renders all pools, and
their statistics. I will reimplement the /resources to return a list
of resources and their information the following commit.