unknown-resources: catch stdout lines from cmd_list securely
The output from cmd_list has a strict semantics; space-separated list of
resource names to be removed potentially.  We can not afford taking a
"part of some stdout line", which could lead even to removal of unwanted
resource.  The same way, we don't want to remove resources with name
'<<trimmed>>'.

While on it, increase the stdout buffer for cmd_list.  We typically have
resource names of length 50 characters.  So provide a space for at least
100 of them.