webui: add a simple homepage
We IMHO don't need to bother with pretty design and content of our
homepage yet. Almost anything is better than printing "OK" :-)

I just wanted to add a simple homepage, so that people can see what is
available, and that we can send a link to just one page, instead of
two (for pools and for resources).