server: stop using sqlalchemy.pool.NullPool
NullPool is automatically used for the sqlite backend anyway, and for
PostgreSQL backend is QueuePool the default.  This patch is solving some
mysterious Manager's hangs in session_scope() queries.

I don't in particular remember why I originally used NullPool, but it
was likely some mistake from early experiments when I started with
sqlite.